+++ plugins/problems/problems.php
@@ -118,6 +118,7 @@ class ProblemsPlugin extends Plugin
         $html = str_replace('%%PROBLEMS%%', $problems, $html);
 
         echo $html;
+        http_response_code(500);
 
         exit();
 
@@ -267,6 +268,19 @@ class ProblemsPlugin extends Plugin
             $mbstring_status = 'error';
         }
         $this->results['mbstring'] = [$mbstring_status => 'PHP Mbstring (Multibyte String Library) is '. $mbstring_adjective . 'installed'];
+
+        // Check Exif if enabled
+        if ($this->grav['config']->get('system.media.auto_metadata_exif')) {
+            if(extension_loaded('exif')) {
+                $exif_adjective = '';
+                $exif_status = 'success';
+            } else {
+                $problems_found = true;
+                $exif_adjective = 'not ';
+                $exif_status = 'error';
+            }
+            $this->results['exif'] = [$exif_status => 'PHP Exif (Exchangeable Image File Format) is '. $exif_adjective . 'installed'];
+        }
 
         // Check for PHP Zip library
         if (extension_loaded('zip')) {