+++ plugins/shortcode-core/classes/ShortcodeManager.php
@@ -4,6 +4,7 @@ namespace Grav\Plugin;
 use Grav\Common\Data\Data;
 use Grav\Common\Grav;
 use Grav\Common\Page\Page;
+use Guzzle\Common\Exception\UnexpectedValueException;
 use Thunder\Shortcode\EventContainer\EventContainer;
 use Thunder\Shortcode\HandlerContainer\HandlerContainer;
 use Thunder\Shortcode\Processor\Processor;
@@ -15,6 +16,9 @@ class ShortcodeManager
 
     /** @var Grav $grav */
     protected $grav;
+
+    /** @var Page $page */
+    protected $page;
 
     /** @var  HandlerContainer $handlers */
     protected $handlers;
@@ -30,6 +34,7 @@ class ShortcodeManager
 
     /**
      * initialize some internal instance variables
+     * @param Page $page
      */
     public function __construct()
     {
@@ -169,11 +174,15 @@ class ShortcodeManager
      */
     public function registerAllShortcodes($directory)
     {
-        foreach (new \DirectoryIterator($directory) as $file) {
-            if($file->isDot()) {
-                continue;
+        try {
+            foreach (new \DirectoryIterator($directory) as $file) {
+                if ($file->isDot()) {
+                    continue;
+                }
+                $this->registerShortcode($file->getFilename(), $directory);
             }
-            $this->registerShortcode($file->getFilename(), $directory);
+        } catch (\UnexpectedValueException $e) {
+            Grav::instance()['log']->error('ShortcodeCore Plugin: Directory not found => ' . $directory);
         }
     }
 
@@ -208,20 +217,10 @@ class ShortcodeManager
      */
     public function processContent(Page $page, Data $config)
     {
-        switch($config->get('parser'))
-        {
-            case 'regular':
-                $parser = 'Thunder\Shortcode\Parser\RegularParser';
-                break;
-            case 'wordpress':
-                $parser = 'Thunder\Shortcode\Parser\WordpressParser';
-                break;
-            default:
-                $parser = 'Thunder\Shortcode\Parser\RegexParser';
-                break;
-        }
+        $parser = $this->getParser($config->get('parser'));
 
         if ($page && $config->get('enabled')) {
+            $this->page = $page;
             $content = $page->getRawContent();
             $processor = new Processor(new $parser(new CommonSyntax()), $this->handlers);
             $processor = $processor->withEventContainer($this->events);
@@ -229,6 +228,22 @@ class ShortcodeManager
 
             return $processed_content;
         }
+    }
+
+    /**
+     * Allow the processing of shortcodes directly on a string
+     * For example when used by Twig directly
+     *
+     * @param $str
+     * @return string
+     */
+    public function processShortcodes($str)
+    {
+        $parser = $this->getParser($this->config->get('parser'));
+        $processor = new Processor(new $parser(new CommonSyntax()), $this->handlers);
+        $processed_string = $processor->process($str);
+
+        return $processed_string;
     }
 
     /**
@@ -264,5 +279,45 @@ class ShortcodeManager
     public function getId(ShortcodeInterface $shortcode)
     {
         return substr(md5($shortcode->getShortcodeText()), -10);
+    }
+
+    /**
+     * Sets the current page context
+     *
+     * @param Page $page
+     */
+    public function setPage(Page $page)
+    {
+        $this->page = $page;
+    }
+
+    /** gets the current page context if set */
+    public function getPage()
+    {
+        return $this->page;
+    }
+
+    /**
+     * Get the appropriate parser object
+     *
+     * @param $parser
+     * @return string
+     */
+    protected function getParser($parser)
+    {
+        switch($parser)
+        {
+            case 'regular':
+                $parser = 'Thunder\Shortcode\Parser\RegularParser';
+                break;
+            case 'wordpress':
+                $parser = 'Thunder\Shortcode\Parser\WordpressParser';
+                break;
+            default:
+                $parser = 'Thunder\Shortcode\Parser\RegexParser';
+                break;
+        }
+
+        return $parser;
     }
 }